SMART Goals and what exactly am I making.
Truth be told, I don't know. And I most certainly don't understand. This is as much about educating myself as it is creating a tool for others to do the same. In the right order, using what they have to make life that little bit more bearable. Put in the hours and you will see the return, and you will have the tools to defend your keep from those who overnight have the technical ability and means to just take it from you. A helping hand in creating something of your own and steer you away from the traps that are getting harder and harder to see. In a format you can not only implement but understand why. A series of not how-tos, but why and what you can expect. Realistic, simple, and educational. I don't know if this is the platform for it either but this is what I have and know, and should some kind hearted souls be interested in educating and correcting the guinea pig so that he can make it clear just how much work needs doing for at first minimal rewards, and like with all things perseverance and work ethic always, always pay off in the end, I and hopefully those that need and want it most can have their yoghurt because the machine they have for Netflix is always on anyway, and their locked savings actually grow more than the rate of inflation and not -4,5% like they would in a standard Swedish locked savings account. It would make me a happy and grateful man. 

That's what all tis was meant to be in the first place I find myself possibly naïvely pondering, a way for the average person to become masters of their own fate. As long as they are prepared for the work it takes to get there.

Smart goal 1. Deadline 16/3.
*Situation: Software lives in hardware. From the ethernet jack to the cellphone, how to affordably secure your keep proactively.
*Measurable? Yes, how exposed will I be a week from now compared to today and how did I do it.
*Attainability? Will this be something Joe can accomplish in the same timeframe?
*Result. A proactive defence, so at least some people don't have to go through the hell that still isn't over for me
*Timeframe. One week. Generous, but that is the concept here, it has to appeal to the masses. And they have to find it reasonable.

All input is welcome, this is still a concept of an idea in the making and my skin is thick so don't hold back. 

Catch you on the flipside!

Raven6345@protonmail.com